requirements.txt
python-telegram-bot==20.7
sentence-transformers==2.2.2
numpy==1.24.3
faiss-cpu==1.7.4
llama-cpp-python==0.2.23
python-dotenv==1.0.0